Phase 14a: extract RNSAnnounceHandler and AnnouncedVersionCache
RNS.java 873 → 713 L. The announce-receive path was a 146-line inner class
holding a quarter of the facade; it is self-contained apart from one Transport
registration, so it lifts out cleanly.

RNSAnnounceHandler is one instance per aspect, as before, with everything that
differed between the two — aspect filter, PeerAspect, peer limit — now a
constructor argument rather than a `QDN_ASPECT.equals(aspectFilter)` test
repeated in two methods. It reads the peer list through the registry it is
given instead of reaching back through RNS.getInstance().

AnnouncedVersionCache is the bounded LRU that was an anonymous LinkedHashMap
field on RNS. It is shared: the handler writes announced versions, the inbound
identify path reads them. A separate class rather than a member of the handler
because both sides need it and neither owns the other.

No behaviour change. The §10 comment (why getAspectFilter() returns null) and
the @Synchronized rationale move verbatim; every other removed comment line is
accounted for in the new files, the version-cache javadoc having grown into
AnnouncedVersionCache's class doc. compile + test-compile clean, 63 tests green.
